refactor: 사진 permission 관련 코드 폴더 변경
